Removed the dependence on the Shapely module.
THe Shapely module was being used for one of the tougher geometry
caluclations, and I had aspirations of using it for all geometry work.
However, since I haven't done so yet, I decided to write that function
myself, in order to remove the dependence.

Making portable code is better
than one super-tested and efficient method.